decision regarding ImageMagic (v6, 16bit, float-HDRI)
Quoting from https://github.com/Ichthyostega/openfx-arena-deb/blob/deb/README.md

> ImageMagick (MagickCore/Magick++) 6.9.2/7.0.3 with Q32, HDRI, lcms2, zlib, freetype, libpng
>  - Will build on older versions, but some features/plugins may not be available
>  - Quantum depth under 32 will work, but it not recommended, note that most distros ship 16
>  - Will work without HDRI, but it's not recommended


However, ImageMagic v7.x is a breaking change and not even Debian/unstable has
started to package it yet. v6.x aka "ImageMagic-legacy" will be supported until 2028

* "quantum depth" defines the scale/range for colour values
  - quantum 8 means 0..255
  - quantum 16 means 0..65535
  - supported values are 8, 16, 32, or 64
* however, the "HDRI" toggle turns all compuation to use floats
  and remove any clamping.
* Thus 16 HDRI means that a colour value of 65535 == 1.0

Since Debian and most other distros ship a compile of ImageMagic with 16+HDRI,
it seems adequate to stick to that. Note also that ImageMagic is highly secuirty relevant
and has had serious vulnerabilities in the past, so I do not want to be
responsible for maintaining a private package.

The relevance of the ImageMagic Natron plug-in is
 - generic I/O of all file formats supported by ImageMagic
 - especially Adobe Photoshop PSD import
 - Image Magic 7.x will be able to import Gimp XCF directly
 - beyond that, the "Magic" subdir provides some further effect filters

Note, Regarding HDR, that Natron uses floats internally for processing
and that there is a separate plug-in for OpenEXR import/export